requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
PyMySQL==1.1.0
cryptography==41.0.4
requests==2.31.0
python-dotenv==1.0.0
APScheduler==3.10.4
textblob==0.17.1
pandas==2.1.1
plotly==5.17.0
Flask-Migrate==4.0.5
Werkzeug==2.3.7
Jinja2==3.1.2
beautifulsoup4==4.12.2
lxml==4.9.3
openpyxl==3.1.2